toys/devops/database_migration/aws_rds_to_s3/requirements.txt
psycopg[binary,pool]
python-dotenv